Phase 2b: fix CI regressions (lint, integration tests, brand-grep)
Three failing jobs on PR #34:

1. build + test + lint (golangci-lint 62 new issues):
   - paralleltest: add t.Parallel() to every new auth test
   - revive: doc comments on CachedStore.{Update,Disable,Change}Role +
     Role const block; rename test local `real` -> `realUser`; drop
     unused *w param in session-middleware test
   - staticcheck QF1008: c.Store.GetUserByID -> c.GetUserByID (embed)
   - unconvert: drop uint32() wrappers on v.GetUint32
   - gocyclo (Login 21>13, TestLogin_HappyPath 14>13): split Login
     into decodeLogin / authenticate / mintSessionAndRespond;
     extract assertSessionCookie helper in tests
   - intrange: modern for-range loops in lockout/session tests
   - dupl: collapse two near-duplicate TestRequireRole tests into a
     table-driven TestRequireRole with sub-tests
   - gofmt / gofumpt: autofix run + combine var() sanity check block
   - errcheck: guard rows.Close() with _ =
   - wsl_v5: blank lines before w.WriteHeader / defer rows.Close

2. Integration tests (test [build failed]):
   - test/dashboard_{http,spa}_test.go referenced the removed
     dashboard.BearerTokenMiddleware; rewire to
     auth.APITokenMiddleware + auth.RequireAuthenticated composite
     (same 401/200 contract, just the new plumbing)

3. Dashboard SPA brand-grep:
   - web/src/App.css + web/src/pages/Login.css used hex fallbacks in
     var(--x, #hex) — repo rule allows hex only in tokens.css.
   - Login.css rewritten to reference real tokens (--paper, --ink,
     --green, --green-dk, --rule-soft, --error, --s-*) with no fallbacks.

Local verification: go test ./... + make smoke (13/13 probes) +
golangci-lint run --new-from-rev=origin/main all green.

+++ cmd/statnive-live/main.go
@@ -682,8 +682,8 @@ type appConfig struct {
 			SameSite   string // "lax" only; other values rejected in v1
 			CacheTTL   time.Duration
 		}
-		BcryptCost      int
-		LoginRateLimit  struct {
+		BcryptCost     int
+		LoginRateLimit struct {
 			Requests int
 			Window   time.Duration
 		}
@@ -813,9 +813,9 @@ func loadConfig() (appConfig, error) {
 	cfg.Auth.Lockout.Fails = v.GetInt("auth.lockout.fails")
 	cfg.Auth.Lockout.Decay = v.GetDuration("auth.lockout.decay")
 	cfg.Auth.Lockout.Lockout = v.GetDuration("auth.lockout.lockout")
-	cfg.Auth.Bootstrap.SiteID = uint32(v.GetUint32("auth.bootstrap.site_id"))
+	cfg.Auth.Bootstrap.SiteID = v.GetUint32("auth.bootstrap.site_id")
 	cfg.Auth.Bootstrap.Username = v.GetString("auth.bootstrap.username")
-	cfg.Auth.DefaultSiteID = uint32(v.GetUint32("auth.default_site_id"))
+	cfg.Auth.DefaultSiteID = v.GetUint32("auth.default_site_id")
 	cfg.Auth.DemoBanner = v.GetString("auth.demo_banner")
 
 	// auth.api_tokens is a list of {token_hash, site_id, label, role}
